requirements.txt
biopython==1.68
ipython==5.2.2
Keras==1.2.2
matplotlib==2.0.0
numpy==1.12.0
pandas==0.19.2
scikit-learn==0.18.1
scipy==0.18.1
six==1.10.0
tensorflow==1.0.0
Theano==0.8.2

